Partner Portal: Move LicenseStates enum to types.ts
