+++ client/jetpack-cloud/sections/partner-portal/license-list/index.tsx
@@ -7,6 +7,7 @@ import { useTranslate } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
+import { LicenseStates } from 'calypso/jetpack-cloud/sections/partner-portal/types';
 import Main from 'calypso/components/main';
 import CardHeading from 'calypso/components/card-heading';
 import DocumentHead from 'calypso/components/data/document-head';
@@ -14,19 +15,15 @@ import LicenseListItem from 'calypso/jetpack-cloud/sections/partner-portal/licen
 import LicensePreview from 'calypso/jetpack-cloud/sections/partner-portal/license-preview';
 import LicenseStateFilter from 'calypso/jetpack-cloud/sections/partner-portal/license-state-filter';
 
-export enum LicenseStates {
-	All = 'all',
-	Detached = 'detached',
-	Attached = 'attached',
-	Revoked = 'revoked',
-}
-
 interface Props {
 	licenseState?: string;
 	search?: string;
 }
 
-export default function LicenseList( { licenseState, search }: Props ): ReactElement {
+export default function LicenseList( {
+	licenseState = LicenseStates.All,
+	search = '',
+}: Props ): ReactElement {
 	const translate = useTranslate();
 	const stateFilter = Object.values( LicenseStates ).includes( licenseState as never )
 		? licenseState
