+++ client/jetpack-cloud/sections/partner-portal/types.ts
@@ -1,0 +1,6 @@
+export enum LicenseStates {
+	All = 'all',
+	Detached = 'detached',
+	Attached = 'attached',
+	Revoked = 'revoked',
+}
